(server/arm): direct servo control will not be handled by the server anymore
the plan is to control the servos by proxy of a Pi Pico that we send data to over a serial connection

+++ shared/src/commonMain/kotlin/me/arianb/storm_robot/SerializableDataClasses.kt
@@ -11,9 +11,9 @@ data class WheelMovementPacket(
 
 @Serializable
 data class ArmMovementPacket(
-    val baseAngle: Int,
-    val shoulderAngle: Int,
-    val elbowAngle: Int,
-    val wristAngle: Int,
-    val gripperAngle: Int
+    val baseAngle: UByte,
+    val shoulderAngle: UByte,
+    val elbowAngle: UByte,
+    val wristAngle: UByte,
+    val gripperAngle: UByte
 )
